Move top-page components from app to components
