+++ components/toppage/Ratings.tsx
@@ -1,5 +1,5 @@
-import Block from "../components/Block";
-import styles from "../styles/top-page/Ratings.module.scss";
+import Block from "../Block";
+import styles from "../../styles/top-page/Ratings.module.scss";
 
 type Props = {
   id?: string
