+++ components/toppage/WhatsNew.tsx
@@ -1,7 +1,7 @@
-import styles from "../styles/top-page/WhatsNew.module.scss";
+import styles from "../../styles/top-page/WhatsNew.module.scss";
 import ReactMarkdown from "react-markdown";
-import Block from "../components/Block";
-import {getWhatsNewRecords, WhatsNewRecord} from "../lib/whats_new";
+import Block from "../Block";
+import {getWhatsNewRecords, WhatsNewRecord} from "../../lib/whats_new";
 import React from "react";
 
 type Props = {
